Work on the slurm distribution
